Updates for Adding Records.
Updated LoginController, changed the fielkd names to the user entity.

@GeneratedValue(strategy=GenerationType.AUTO)  added to all entities, except M:M relations.

Updated /Login/Register to include additional attributes.

Commented out password check for testing

+++ src/main/java/org/fyp/controller/UserController.java
@@ -1,6 +1,7 @@
 package org.fyp.controller;
 
 import org.fyp.model.*;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.text.ParseException;
 import java.util.Collection;
+import java.util.HashMap;
 
 /**
  * Created by oisin on 18/07/2017.
@@ -35,22 +37,76 @@ public class UserController extends MainController {
         userRepo.save(user);
     }
 
-    @RequestMapping(value = "/delete", method=RequestMethod.GET)
-    public Collection<User> delete(User user) {
-        userRepo.delete(user);
-        return userRepo.findAll();
+    @RequestMapping(value = "/delete", method=RequestMethod.DELETE)
+    public ResponseEntity<HashMap<String,String>> delete(User user) {
+
+        respMap.put("httpStatus",""+httpStatus);
+        // check if user exists
+        // delete
+        // check if user has been deleted.
+
+        if (userRepo.findByUserid(user.getUserid()) != null) {
+            // delete user
+            userRepo.delete(user);
+            // check user has been deleted
+            if (userRepo.findByUserid(user.getUserid()) == null) {
+                // return success
+                respMap.put("message","User deleted successfully");
+                respMap.put("success","1");
+            } else {
+                // return failure
+                respMap.put("message","User could not be deleted");
+                respMap.put("success","0");
+            }
+
+        } else {
+            respMap.put("message","User does not exist in database");
+            respMap.put("success","0");
+            // return user does not exist in db
+        }
+
+        return  new ResponseEntity<>(respMap, httpStatus);
     }
 
-    @RequestMapping(value = "/deleteByEmail/{email}", method=RequestMethod.GET)
-    public Collection<User> delete(@PathVariable ("email") String email) {
-        userRepo.deleteByEmailAddress(email);
-        return userRepo.findAll();
+    @RequestMapping(value = "/deleteByEmail/{email:.+}", method=RequestMethod.GET)
+    public ResponseEntity<HashMap<String,String>> delete(@PathVariable ("email") String email) {
+
+        if (userRepo.findByEmailAddress(email) != null) {
+            // delete user
+            userRepo.deleteByEmailAddress(email);
+            // check user has been deleted
+            if (userRepo.findByEmailAddress(email) == null) {
+                // return success
+                respMap.put("message","User deleted successfully");
+                respMap.put("success","1");
+            } else {
+                // return failure
+                respMap.put("message","User could not be deleted By Email Address");
+                respMap.put("success","0");
+            }
+
+        } else {
+            respMap.put("message","User does not exist in database");
+            respMap.put("success","0");
+            // return user does not exist in db
+        }
+        return new ResponseEntity<>(respMap, httpStatus);
     }
 
     @RequestMapping(value = "/delete/{id}", method=RequestMethod.GET)
-    public Collection<User> delete(@PathVariable ("id") int id) {
-        userRepo.deleteByUserid(id);
-        return userRepo.findAll();
+    public ResponseEntity<HashMap<String,String>> delete(@PathVariable ("id") int id) {
+
+        if( userRepo.deleteByUserid(id) > 0) {
+            // return success
+            respMap.put("message","User deleted successfully");
+            respMap.put("success","1");
+        } else {
+            // return failure
+            respMap.put("message","User could not be deleted");
+            respMap.put("success","0");
+        }
+
+        return new ResponseEntity<>(respMap, httpStatus);
     }
 
     // the pathmapping variables are truncated after the '.', ie 'abc@dd.com' is truncated to 'abc@dd.'
@@ -64,6 +120,24 @@ public class UserController extends MainController {
     public User getUserBy(@PathVariable("id") int id)
     {
         return userRepo.findByUserid(id);
+    }
+
+
+    @RequestMapping(value = {"/activate/{id}"}, method=RequestMethod.GET)
+    public ResponseEntity<HashMap<String,String>> activateUser(@PathVariable("id") int id)
+    {
+        User user = userRepo.findByUserid(id);
+        if (user != null) {
+            user.setActive((byte)1);
+            respMap.put("message","User:" + user.buildFullname() + " is activated");
+            respMap.put("success","1");
+        } else {
+            // return failure
+            respMap.put("message","User could not be activated");
+            respMap.put("success","0");
+        }
+
+        return new ResponseEntity<>(respMap, httpStatus);
     }
